third_party/WebKit/LayoutTests/webaudio/offlineaudiocontext-suspend-resume-sequence-expected.txt

Test OfflineAudioContext.resume() and OfflineAudioContext.suspend() with the timed sequence.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The actual resolution order of the suspend 5 (0.00 sec.) is equal to 0.
PASS The actual resolution order of the suspend 0 (0.25 sec.) is equal to 1.
PASS The actual resolution order of the suspend 3 (0.50 sec.) is equal to 2.
PASS The actual resolution order of the suspend 1 (0.75 sec.) is equal to 3.
PASS The actual resolution order of the suspend 2 (1.00 sec.) is equal to 4.
PASS The actual resolution order of the suspend 4 (1.25 sec.) is equal to 5.
PASS The actual resolution order of the suspend 6 (1.75 sec.) is equal to 6.
PASS successfullyParsed is true

TEST COMPLETE
